+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## 0.1.0 \(YYYY-MM-DD\)
 
 ### New Features
+* Add NerdDice.total_dice class method
 
 ### Bug Fixes